+++ ttl_dict/_exceptions.py
@@ -1,6 +1,0 @@
-class TTLDictError(Exception):
-    pass
-
-
-class TTLDictInvalidConfigError(TTLDictError):
-    pass